backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
scikit-learn==1.4.1.post1
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.6.3
websockets==12.0
requests==2.31.0
google-generativeai==0.4.1
